+++ jest.setup.js
@@ -1,6 +1,27 @@
 import '@testing-library/jest-dom/extend-expect';
 import { server } from './__mocks__/server.js';
 
+// Mock IntersectionObserver
+global.IntersectionObserver = class IntersectionObserver {
+    constructor() { }
+
+    disconnect() {
+        return null;
+    }
+
+    observe() {
+        return null;
+    }
+
+    takeRecords() {
+        return null;
+    }
+
+    unobserve() {
+        return null;
+    }
+};
+
 // Establish API mocking before all tests.
 beforeAll(() => server.listen())
 // Reset any request handlers that we may add during the tests,