+++ src/Components/Cards/index.js
@@ -1,0 +1,7 @@
+import AppCard from './AppCard/AppCard';
+import PluginCard from './PluginCard/PluginCard';
+
+export {
+    AppCard,
+    PluginCard,
+}